moved assignment class to Posttype folder
