+++ source/tests/php/AssignmentTest.php
@@ -4,7 +4,7 @@ namespace php;
 
 use Brain\Monkey\Functions;
 use PluginTestCase\PluginTestCase;
-use VolunteerManager\Assignment;
+use VolunteerManager\PostType\Assignment\Assignment;
 
 class AssignmentTest extends PluginTestCase
 {
